// temper/shared-addon-libraries-lib-data-encode/src/types.ts
export type LdeValue = string | number

export type LuaTable = Record<string | number, unknown>

export type GlobalTable = Record<string, unknown>

export type LuaArray = unknown[]

export type OptionalNumber = number | undefined

export type LdeValueArray = LdeValue[]

export interface TestResult {
  testDictGlobal?: LdeValue[]
  encoded?: string[]
  decoded?: unknown
  dict?: LdeValue[]
  result?: boolean
  encoder?: EncodeInstance
  decoder?: DecodeInstance
}

export interface DictionaryInstance {
  globalDictReverse: LuaTable
  dictionary: LdeValue[]
  counts: [LuaTable, LuaTable, LuaTable]
  Initialize: (this: DictionaryInstance, data: unknown, globalDictionary?: LdeValue[]) => void
  ScanTable: (this: DictionaryInstance, data: unknown) => void
  ValidateValue: (this: DictionaryInstance, value: unknown) => boolean
  IncreaseCount: (this: DictionaryInstance, value: LdeValue) => void
  [key: string]: unknown
}

export interface DictionaryClass {
  Subclass: (this: DictionaryClass) => DictionaryClass
  New: (this: DictionaryClass, data: unknown, globalDictionary?: LdeValue[]) => DictionaryInstance
  Initialize: (this: DictionaryInstance, data: unknown, globalDictionary?: LdeValue[]) => void
  ScanTable: (this: DictionaryInstance, data: unknown) => void
  ValidateValue: (this: DictionaryInstance, value: unknown) => boolean
  IncreaseCount: (this: DictionaryInstance, value: LdeValue) => void
  [key: string]: unknown
}